Move helper function to cmdutil.
